Publishing Last In-Dev Build, Now in Testing.
Controls (for the Player and Menus) and level design are final. Will be playtesting on Base PS4 and PS4 pro, to test for anomalies in performance. Anything that can speed up performance will be implemented. Minor frustrations and pain points will be tweaked as well, but those should be ironed out.